requirements.txt
Django==3.0.9
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
pyjwt==1.7.1
pylint==2.5.3
autopep8==1.5.4
python-dotenv==0.14.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
django-heroku==0.3.1
